requirements.txt
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-binary==128.0.6576.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.3.0
Django==4.2.13
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.16.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.53.1
fqdn==1.5.1
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pbr==6.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.1.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.21.12
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
scikit-learn==1.6.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.22.0
Send2Trash==1.8.3
setuptools==80.0.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.2
stevedore==5.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.0.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.9.0
widgetsnbextension==4.0.11
wsproto==1.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Test Data Generation
factory-boy==3.3.0
faker==20.1.0
